requirements.txt
streamlit
tensorflow
torch
transformers
pillow
midiutil
pretty_midi
pydub
